feat: duration serializing
