+++ src/duration/mod.rs
@@ -1,4 +1,5 @@
 pub mod format;
 pub mod relative;
+pub mod serde;
 
 pub use relative::*;
